-Added the ability to change user roles via the admin panel -Added the main admin page, which displays site statistics: number of registered users, number of active and completed tickets, as well as the number of unassigned tickets. From the main admin page, you can now go to the user list page and the ticket list page -A separate page has been added for editing tickets, with detailed information about the ticket and the ability to edit, delete, change its status, and assign an executor -Added the ability to delete a ticket and change its status -An interface has been added for almost every page

+++ src/main/java/org/rockend/ticket_system/dto/ChangeUserRoleDto.java
@@ -1,4 +1,4 @@
 package org.rockend.ticket_system.dto;
 
 
-public record ChangeUserRoleDto(int id, String newRole) { }
+public record ChangeUserRoleDto(long id, String newRole) { }
